requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
transformers==4.33.2
torch>=2.2.0
pydantic==2.3.0
python-dotenv==1.0.0
hazm==0.7.0
